Deduplicate materialization event dispatch
